+++ src/Yasumi/Provider/Australia.php
@@ -14,9 +14,9 @@ namespace Yasumi\Provider;
 
 use DateInterval;
 use DateTime;
-use Yasumi\Exception\InvalidDateException;
 use Yasumi\Exception\UnknownLocaleException;
 use Yasumi\Holiday;
+use Yasumi\SubstituteHoliday;
 
 /**
  * Provider for all holidays in Australia.
@@ -69,49 +69,36 @@ class Australia extends AbstractProvider
      */
     private function calculateNewYearHolidays(): void
     {
-        $newyearsday = new DateTime("$this->year-01-01", DateTimeZoneFactory::getDateTimeZone($this->timezone));
-        $this->calculateHoliday('newYearsDay', $newyearsday, [], false, false);
-        switch ($newyearsday->format('w')) {
+        $newYearsDay = new DateTime("$this->year-01-01", DateTimeZoneFactory::getDateTimeZone($this->timezone));
+        $this->addHoliday(new Holiday(
+            'newYearsDay',
+            [],
+            $newYearsDay,
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
+        switch ($newYearsDay->format('w')) {
             case 0: // sunday
-                $newyearsday->add(new DateInterval('P1D'));
-                $this->calculateHoliday('newYearsHoliday', $newyearsday, ['en' => 'New Year’s Holiday'], false, false);
+                $newYearsDay->add(new DateInterval('P1D'));
+                $this->addHoliday(new Holiday(
+                    'newYearsHoliday',
+                    ['en' => 'New Year’s Holiday'],
+                    $newYearsDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
                 break;
             case 6: // saturday
-                $newyearsday->add(new DateInterval('P2D'));
-                $this->calculateHoliday('newYearsHoliday', $newyearsday, ['en' => 'New Year’s Holiday'], false, false);
+                $newYearsDay->add(new DateInterval('P2D'));
+                $this->addHoliday(new Holiday(
+                    'newYearsHoliday',
+                    ['en' => 'New Year’s Holiday'],
+                    $newYearsDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
                 break;
         }
-    }
-
-    /**
-     * Function to simplify moving holidays to mondays if required
-     *
-     * @param string $shortName
-     * @param DateTime $date
-     * @param array $names
-     * @param bool $moveFromSaturday
-     * @param bool $moveFromSunday
-     * @param string $type
-     *
-     * @throws InvalidDateException
-     * @throws \InvalidArgumentException
-     * @throws UnknownLocaleException
-     * @throws \Exception
-     */
-    public function calculateHoliday(
-        string $shortName,
-        DateTime $date,
-        array $names = [],
-        ?bool $moveFromSaturday = null,
-        ?bool $moveFromSunday = null,
-        ?string $type = null
-    ): void {
-        $day = (int)$date->format('w');
-        if ((0 === $day && ($moveFromSunday ?? true)) || (6 === $day && ($moveFromSaturday ?? true))) {
-            $date = $date->add(0 === $day ? new DateInterval('P1D') : new DateInterval('P2D'));
-        }
-
-        $this->addHoliday(new Holiday($shortName, $names, $date, $this->locale, $type ?? Holiday::TYPE_OFFICIAL));
     }
 
     /**
@@ -136,7 +123,27 @@ class Australia extends AbstractProvider
     {
         $date = new DateTime("$this->year-01-26", DateTimeZoneFactory::getDateTimeZone($this->timezone));
 
-        $this->calculateHoliday('australiaDay', $date, ['en' => 'Australia Day']);
+        $holiday = new Holiday(
+            'australiaDay',
+            ['en' => 'Australia Day'],
+            $date,
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        );
+        $this->addHoliday($holiday);
+
+        $day = (int)$date->format('w');
+        if (0 === $day || 6 === $day) {
+            $date = $date->add(0 === $day ? new DateInterval('P1D') : new DateInterval('P2D'));
+
+            $this->addHoliday(new SubstituteHoliday(
+                $holiday,
+                [],
+                $date,
+                $this->locale,
+                Holiday::TYPE_OFFICIAL
+            ));
+        }
     }
 
     /**
@@ -162,7 +169,13 @@ class Australia extends AbstractProvider
         }
 
         $date = new DateTime("$this->year-04-25", DateTimeZoneFactory::getDateTimeZone($this->timezone));
-        $this->calculateHoliday('anzacDay', $date, [], false, false);
+        $this->addHoliday(new Holiday(
+            'anzacDay',
+            [],
+            $date,
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
         $easter = $this->calculateEaster($this->year, $this->timezone);
 
         $easterMonday = $this->calculateEaster($this->year, $this->timezone);
@@ -171,9 +184,14 @@ class Australia extends AbstractProvider
         $fDate = $date->format('Y-m-d');
         if ($fDate === $easter->format('Y-m-d') || $fDate === $easterMonday->format('Y-m-d')) {
             $easterMonday->add(new DateInterval('P1D'));
-            $this->calculateHoliday('easterTuesday', $easterMonday, ['en' => 'Easter Tuesday'], false, false);
+            $this->addHoliday(new Holiday(
+                'easterTuesday',
+                ['en' => 'Easter Tuesday'],
+                $easterMonday,
+                $this->locale,
+                Holiday::TYPE_OFFICIAL
+            ));
         }
-        unset($fDate);
     }
 
     /**
@@ -192,23 +210,59 @@ class Australia extends AbstractProvider
     {
         $christmasDay = new DateTime("$this->year-12-25", DateTimeZoneFactory::getDateTimeZone($this->timezone));
         $boxingDay = new DateTime("$this->year-12-26", DateTimeZoneFactory::getDateTimeZone($this->timezone));
-        $this->calculateHoliday('christmasDay', $christmasDay, [], false, false);
-        $this->calculateHoliday('secondChristmasDay', $boxingDay, [], false, false);
+        $this->addHoliday(new Holiday(
+            'christmasDay',
+            [],
+            $christmasDay,
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
+        $this->addHoliday(new Holiday(
+            'secondChristmasDay',
+            [],
+            $boxingDay,
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
 
         switch ($christmasDay->format('w')) {
             case 0: // sunday
                 $christmasDay->add(new DateInterval('P2D'));
-                $this->calculateHoliday('christmasHoliday', $christmasDay, ['en' => 'Christmas Holiday'], false, false);
+                $this->addHoliday(new Holiday(
+                    'christmasHoliday',
+                    ['en' => 'Christmas Holiday'],
+                    $christmasDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
                 break;
             case 5: // friday
                 $boxingDay->add(new DateInterval('P2D'));
-                $this->calculateHoliday('secondChristmasHoliday', $boxingDay, ['en' => 'Boxing Day Holiday'], false, false);
+                $this->addHoliday(new Holiday(
+                    'secondChristmasHoliday',
+                    ['en' => 'Boxing Day Holiday'],
+                    $boxingDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
                 break;
             case 6: // saturday
                 $christmasDay->add(new DateInterval('P2D'));
                 $boxingDay->add(new DateInterval('P2D'));
-                $this->calculateHoliday('christmasHoliday', $christmasDay, ['en' => 'Christmas Holiday'], false, false);
-                $this->calculateHoliday('secondChristmasHoliday', $boxingDay, ['en' => 'Boxing Day Holiday'], false, false);
+                $this->addHoliday(new Holiday(
+                    'christmasHoliday',
+                    ['en' => 'Christmas Holiday'],
+                    $christmasDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
+                $this->addHoliday(new Holiday(
+                    'secondChristmasHoliday',
+                    ['en' => 'Boxing Day Holiday'],
+                    $boxingDay,
+                    $this->locale,
+                    Holiday::TYPE_OFFICIAL
+                ));
                 break;
         }
     }