+++ src/Yasumi/Provider/Australia/NSW.php
@@ -104,13 +104,13 @@ class NSW extends Australia
      */
     private function calculateQueensBirthday(): void
     {
-        $this->calculateHoliday(
+        $this->addHoliday(new Holiday(
             'queensBirthday',
-            new DateTime('second monday of june ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
             [],
-            false,
-            false
-        );
+            new DateTime('second monday of june ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
     }
 
     /**
@@ -133,13 +133,12 @@ class NSW extends Australia
      */
     private function calculateBankHoliday(): void
     {
-        $this->calculateHoliday(
+        $this->addHoliday(new Holiday(
             'bankHoliday',
-            new DateTime('first monday of august ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
             ['en' => 'Bank Holiday'],
-            false,
-            false,
+            new DateTime('first monday of august ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $this->locale,
             Holiday::TYPE_BANK
-        );
+        ));
     }
 }