+++ src/Yasumi/Provider/Australia/Tasmania/KingIsland.php
@@ -14,6 +14,7 @@ namespace Yasumi\Provider\Australia\Tasmania;
 
 use DateTime;
 use Yasumi\Exception\UnknownLocaleException;
+use Yasumi\Holiday;
 use Yasumi\Provider\Australia\Tasmania;
 use Yasumi\Provider\DateTimeZoneFactory;
 
@@ -51,12 +52,12 @@ class KingIsland extends Tasmania
      */
     private function calculateKingIslandShow(): void
     {
-        $this->calculateHoliday(
+        $this->addHoliday(new Holiday(
             'kingIslandShow',
-            new DateTime('first tuesday of march ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
             ['en' => 'King Island Show'],
-            false,
-            false
-        );
+            new DateTime('first tuesday of march ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
     }
 }