+++ src/Yasumi/Provider/Australia/Tasmania/South/Southeast.php
@@ -14,6 +14,7 @@ namespace Yasumi\Provider\Australia\Tasmania\South;
 
 use DateTime;
 use Yasumi\Exception\UnknownLocaleException;
+use Yasumi\Holiday;
 use Yasumi\Provider\Australia\Tasmania\South;
 use Yasumi\Provider\DateTimeZoneFactory;
 
@@ -54,12 +55,12 @@ class Southeast extends South
      */
     private function calculateHobartRegatta(): void
     {
-        $this->calculateHoliday(
+        $this->addHoliday(new Holiday(
             'hobartRegatta',
-            new DateTime('second monday of february ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
             ['en' => 'Royal Hobart Regatta'],
-            false,
-            false
-        );
+            new DateTime('second monday of february ' . $this->year, DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $this->locale,
+            Holiday::TYPE_OFFICIAL
+        ));
     }
 }